requirements.txt
torch
torchvision
pandas
pillow
tqdm
fastapi
uvicorn
python-multipart
boto3